Rename saved query `group_bys` to `group_by`
In https://github.com/dbt-labs/dbt-semantic-interfaces/pull/186 we
renamed `group_bys` to `group_by` in DSI (which went out in DSI 0.4.0).
The original `group_bys` naming was a typo. This fixes that. This is a
non-breaking change from the core perspective because saved queries
are being introduced in 1.7.0, which has not been cut yet.